Simplifies by removing Structs (just classes)
