+++ src/nodes/ClassNode.cpp
@@ -15,13 +15,12 @@
 
 
 ClassNode::ClassNode(std::string typeName_, std::string name_)
-    : StructNode(std::move(typeName_), std::move(name_))
+    : typeName(std::move(typeName_)), name(std::move(name_))
 {
 }
 
 AnyObject::Ptr ClassNode::evaluate(Scope &scope)
 {
-    // TODO: - inefficient, should have another method we can call to do most of StructNode::evaluate.
     if (active)
     {
         ThrowException("ClassNode named " + name + " of type " + typeName + " is already active");
@@ -32,10 +31,9 @@ AnyObject::Ptr ClassNode::evaluate(Scope &scope)
     // Initialize our instance from the struct definition defined in the scope.
     auto theObject = scope.getNamedObject(typeName);
 
-    structDefinition = std::static_pointer_cast<StructDefinitionNode>(theObject->getValue<BaseNode::Ptr>());
-    structDefinition->installVariablesInScope(_instanceScope, variableNames);
+    classDefinition = std::static_pointer_cast<ClassDefinitionNode>(theObject->getValue<BaseNode::Ptr>());
 
-    auto classDefinition = std::static_pointer_cast<ClassDefinitionNode>(structDefinition);
+    classDefinition->installVariablesInScope(_instanceScope, variableNames);
     classDefinition->installMethodsInScope(_instanceScope);
 
     // Add the active struct instance to the scope. TODO: - transfer ownership
@@ -44,4 +42,31 @@ AnyObject::Ptr ClassNode::evaluate(Scope &scope)
 
     scope.linkObject(name, wrappedClass);
     return wrappedClass;
+}
+
+
+ClassNode &ClassNode::operator=(const ClassNode &other)
+{
+    if (this == &other)
+    {
+        return (*this);
+    }
+
+    // 1. check that both are instances of the same struct type.
+    if (other.classDefinition != classDefinition)
+    {
+        ThrowException("cannot assign Class objects with different definitions");
+    }
+
+    // 2. iterate over the objects stored in the scopes and assign. // TODO: - make more efficient
+    for (auto &variableName : variableNames)
+    {
+        AnyObject::Ptr thisObject = _instanceScope.getNamedObject(variableName);
+        AnyObject::Ptr otherObject = other._instanceScope.getNamedObject(variableName);
+
+        // Attempt an assignment. Will fail if different types.
+        (*thisObject) = (*otherObject);
+    }
+
+    return (*this);
 }
