+++ src/nodes/ClassNode.hpp
@@ -10,17 +10,23 @@
 #pragma once
 #include "ClassDefinitionNode.hpp"
 #include "Exceptions.hpp"
-#include "StructNode.hpp"
 #include <memory>
 #include <string>
+
 
 /**
  * An instance of a class defined by the ClassDefinitionNode.
  */
-class ClassNode : public StructNode
+class ClassNode : public BaseNode, public std::enable_shared_from_this<ClassNode>
 {
 public:
+    using Ptr = std::shared_ptr<ClassNode>;
+
+    ClassNode() = delete;
     ClassNode(std::string typeName_, std::string name_);
+
+    ClassNode &operator=(const ClassNode &other);
+
 
     // ClassNode &operator=(const BaseObject &) override
     // {
@@ -33,4 +39,37 @@ public:
      * @return BaseObject* Pointer to itself
      */
     std::shared_ptr<class AnyObject> evaluate(Scope &scope) override;
+
+    [[nodiscard]] const Scope &instanceScope() const { return _instanceScope; }
+
+    [[nodiscard]] Scope &instanceScope() { return _instanceScope; }
+
+protected:
+    /**
+     * The struct has its own scope for storing its own variables. It does not
+     * inherit from any parent scopes. Literally just used for storing stuff.
+     * When this instance goes out of scope, all variables will be deleted.
+     */
+    Scope _instanceScope;
+
+    /**
+     * Stores names of all variables stored in set.
+     */
+    std::unordered_set<std::string> variableNames;
+
+    /**
+     * Ony a single evaluate call allowed.
+     */
+    bool active{false};
+
+    /**
+     * Stores name of instance and the type.
+     */
+    std::string typeName;
+    std::string name;
+
+    /**
+     * Store the struct definition once active.
+     */
+    std::shared_ptr<ClassDefinitionNode> classDefinition{nullptr};
 };
